+++ smartcontract/programs/doublezero-serviceability/src/processors/accesspass/close.rs
@@ -73,7 +73,7 @@ pub fn process_close_access_pass(
     // Parse the global state account & check if the payer is in the allowlist
     let globalstate = GlobalState::try_from(globalstate_account)?;
     if !globalstate.foundation_allowlist.contains(payer_account.key)
-        && globalstate.reservation_authority_pk != *payer_account.key
+        && globalstate.feed_authority_pk != *payer_account.key
     {
         return Err(DoubleZeroError::NotAllowed.into());
     }
@@ -86,11 +86,11 @@ pub fn process_close_access_pass(
         }
         let accesspass = AccessPass::try_from(accesspass_account)?;
 
-        // Reservation authority can only close access passes they own
-        if globalstate.reservation_authority_pk == *payer_account.key
+        // Feed authority can only close access passes they own
+        if globalstate.feed_authority_pk == *payer_account.key
             && accesspass.owner != *payer_account.key
         {
-            msg!("Reservation authority can only close access passes they own");
+            msg!("Feed authority can only close access passes they own");
             return Err(DoubleZeroError::NotAllowed.into());
         }