+++ smartcontract/sdk/go/serviceability/state.go
@@ -1069,7 +1069,7 @@ const (
 	PermissionFlagNetworkAdmin     uint64 = 1 << 3
 	PermissionFlagTenantAdmin      uint64 = 1 << 4
 	PermissionFlagMulticastAdmin   uint64 = 1 << 5
-	PermissionFlagReservation      uint64 = 1 << 6
+	PermissionFlagFeedAuthority    uint64 = 1 << 6
 	PermissionFlagActivator        uint64 = 1 << 7
 	PermissionFlagSentinel         uint64 = 1 << 8
 	PermissionFlagUserAdmin        uint64 = 1 << 9