+++ smartcontract/sdk/rs/src/commands/user/update.rs
@@ -226,7 +226,7 @@ mod tests {
             health_oracle_pk: Pubkey::new_unique(),
             qa_allowlist: vec![],
             feature_flags: FeatureFlag::OnChainAllocation.to_mask(),
-            reservation_authority_pk: Pubkey::default(),
+            feed_authority_pk: Pubkey::default(),
         };
         client
             .expect_get()
@@ -349,7 +349,7 @@ mod tests {
             health_oracle_pk: Pubkey::new_unique(),
             qa_allowlist: vec![],
             feature_flags: FeatureFlag::OnChainAllocation.to_mask(),
-            reservation_authority_pk: Pubkey::default(),
+            feed_authority_pk: Pubkey::default(),
         };
         client
             .expect_get()